academic_agent.py
from typing import Annotated, TypedDict, Sequence
from langgraph.graph import Graph, StateGraph
from langchain_core.messages import HumanMessage, AIMessage
from langchain_openai import ChatOpenAI
from pydantic import BaseModel
import os
from dotenv import load_dotenv

# 加载环境变量
load_dotenv()

# 定义状态类型
class AgentState(TypedDict):
    messages: Sequence[HumanMessage | AIMessage]
    task_type: str | None
    literature_results: list | None
    summary: str | None
    citations: list | None
    analysis_results: dict | None

# 定义各个节点
def search_literature(state: AgentState) -> AgentState:
    """文献检索节点"""
    # TODO: 实现文献检索逻辑
    return state

def summarize_and_explain(state: AgentState) -> AgentState:
    """文献摘要与术语解释节点"""
    # TODO: 实现摘要和解释逻辑
    return state

def check_citation_validity(state: AgentState) -> AgentState:
    """引用与元数据验证节点"""
    # TODO: 实现引用验证逻辑
    return state

def polish_writing(state: AgentState) -> AgentState:
    """语言提升与翻译节点"""
    # TODO: 实现写作润色逻辑
    return state

def analyze_data(state: AgentState) -> AgentState:
    """数据解析节点"""
    # TODO: 实现数据分析逻辑
    return state

def generate_references(state: AgentState) -> AgentState:
    """格式化引用节点"""
    # TODO: 实现引用生成逻辑
    return state

# 创建工作流图
def create_academic_workflow() -> Graph:
    # 创建工作流
    workflow = StateGraph(AgentState)
    
    # 添加节点
    workflow.add_node("search_literature", search_literature)
    workflow.add_node("summarize_and_explain", summarize_and_explain)
    workflow.add_node("check_citation_validity", check_citation_validity)
    workflow.add_node("polish_writing", polish_writing)
    workflow.add_node("analyze_data", analyze_data)
    workflow.add_node("generate_references", generate_references)
    
    # 设置边和条件
    workflow.add_edge("search_literature", "summarize_and_explain")
    workflow.add_edge("summarize_and_explain", "check_citation_validity")
    
    # 添加条件分支
    def route_by_task_type(state: AgentState) -> str:
        if state["task_type"] == "writing":
            return "polish_writing"
        elif state["task_type"] == "analysis":
            return "analyze_data"
        else:
            return "generate_references"
    
    workflow.add_conditional_edges(
        "check_citation_validity",
        route_by_task_type,
        {
            "polish_writing": "polish_writing",
            "analyze_data": "analyze_data",
            "generate_references": "generate_references"
        }
    )
    
    # 设置入口和出口
    workflow.set_entry_point("search_literature")
    workflow.set_finish_point("polish_writing")
    workflow.set_finish_point("analyze_data")
    workflow.set_finish_point("generate_references")
    
    return workflow.compile()

# 主函数
def main():
    # 创建工作流实例
    workflow = create_academic_workflow()
    
    # 初始化状态
    initial_state = {
        "messages": [],
        "task_type": None,
        "literature_results": None,
        "summary": None,
        "citations": None,
        "analysis_results": None
    }
    
    # 运行工作流
    result = workflow.invoke(initial_state)
    return result

if __name__ == "__main__":
    main() 