requirements.txt
langgraph>=0.0.15
langchain>=0.1.0
langchain-openai>=0.0.2
pydantic>=2.0.0
python-dotenv>=1.0.0
scholarly>=1.7.11
pandas>=2.0.0
numpy>=1.24.0
dashscope>=1.13.6  # Qwen API 客户端 